testing/web-platform/tests/workers/semantics/multiple-workers/006.html



















<!doctype html>
<title>shared worker in dedicated worker in shared worker</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var w1 = new SharedWorker('#1');
  w1.port.onmessage = this.step_func(function(e) {
    assert_equals(e.data, '123');
    this.done();
  });
});
</script>
