third_party/blink/web_tests/platform/generic/external/wpt/workers/semantics/multiple-workers/006-expected.txt

This is a testharness.js-based test.
FAIL shared worker in dedicated worker in shared worker assert_equals: expected "123" but got "Uncaught ReferenceError: Worker is not defined"
Harness: the test ran to completion.
